Cambios para el test corredor y cambio de atributo de Integer a Double

+++ src/ar/edu/unlam/pb2/smartwatch/sensor/Brazada.java
@@ -4,7 +4,7 @@ public class Brazada extends Sensor {
 	private static final Integer LONGITUD_DE_BRAZADA = 1;
 	private Integer contBrazada;
 
-	public Brazada(Integer frecuenciaCardiaca) {
+	public Brazada(Double frecuenciaCardiaca) {
 		super(frecuenciaCardiaca);
 		this.contBrazada = 0;
 	}
